Add watch without selection (#171449)
* 🎁 Support adding expression to watch list without text selection


* 🔨 Remove non-empty selection from "Add to Watch" pre-conditions


* 🔨 Refactor duplicate expression extraction methods into a util function


* 🔨 Make cancellation token parameter optional


* 🔨 Omit unnecessary cancellation token argument


* 🔥 Remove unused import
